+++ src/by/it/group310902/kasperets/lesson01/FiboA.java
@@ -15,6 +15,7 @@ public class FiboA {
         FiboA fibo = new FiboA();
         int n = 33;
         System.out.printf("calc(%d)=%d \n\t time=%d \n\n", n, fibo.calc(n), fibo.time());
+
         //вычисление чисел фибоначчи медленным методом (рекурсией)
         fibo = new FiboA();
         n = 34;
@@ -31,12 +32,11 @@ public class FiboA {
         //здесь простейший вариант, в котором код совпадает
         //с математическим определением чисел Фибоначчи
         //время O(2^n)
-        if (n <= 1) {
-            return n;
-        } else {
-            return calc(n-1) + calc(n-2);
-        }
 
+        if (n < 2) {
+            return n;
+        }
+        return calc(n - 1) + calc(n - 2);
     }
 
 
@@ -45,12 +45,13 @@ public class FiboA {
         //здесь нужно реализовать вариант без ограничения на размер числа,
         //в котором код совпадает с математическим определением чисел Фибоначчи
         //время O(2^n)
-        if (n <= 1) {
-            return BigInteger.valueOf(n);
-        } else {
-            return slowA(n-1).add(slowA(n-2));
+        if (n == 0) {
+            return BigInteger.ZERO;
         }
-
+        if (n == 1) {
+            return BigInteger.ONE;
+        }
+        return slowA(n - 1).add(slowA(n - 2));
 
     }
