+++ src/by/it/group310902/kasperets/lesson04/Lesson4Test.java
@@ -1,4 +1,4 @@
-package by.it.group310901.usov.lesson04;
+package by.it.group310902.kasperets.lesson04;
 
 import org.junit.Test;
