+++ src/by/it/group310902/kasperets/lesson04/dataA.txt
