+++ src/by/it/group310902/kasperets/lesson05/C_QSortOptimized.java
@@ -1,4 +1,4 @@
-package by.it.group310901.usov.lesson05;
+package by.it.group310902.kasperets.lesson05;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -6,25 +6,19 @@ import java.io.InputStream;
 import java.util.Scanner;
 
 /*
-Видеорегистраторы и площадь.
-На площади установлена одна или несколько камер.
-Известны данные о том, когда каждая из них включалась и выключалась (отрезки работы)
-Известен список событий на площади (время начала каждого события).
-Вам необходимо определить для каждого события сколько камер его записали.
+Видеорегистраторы и площадь 2.
+Условие то же что и в задаче А.
 
-В первой строке задано два целых числа:
-    число включений камер (отрезки) 1<=n<=50000
-    число событий (точки) 1<=m<=50000.
+        По сравнению с задачей A доработайте алгоритм так, чтобы
+        1) он оптимально использовал время и память:
+            - за стек отвечает элиминация хвостовой рекурсии,
+            - за сам массив отрезков - сортировка на месте
+            - рекурсивные вызовы должны проводиться на основе 3-разбиения
 
-Следующие n строк содержат по два целых числа ai и bi (ai<=bi) -
-координаты концов отрезков (время работы одной какой-то камеры).
-Последняя строка содержит m целых чисел - координаты точек.
-Все координаты не превышают 10E8 по модулю (!).
+        2) при поиске подходящих отрезков для точки реализуйте метод бинарного поиска
+        для первого отрезка решения, а затем найдите оставшуюся часть решения
+        (т.е. отрезков, подходящих для точки, может быть много)
 
-Точка считается принадлежащей отрезку, если она находится внутри него или на границе.
-
-Для каждой точки в порядке их появления во вводе выведите,
-скольким отрезкам она принадлежит.
     Sample Input:
     2 3
     0 5
@@ -35,13 +29,16 @@ import java.util.Scanner;
 
 */
 
-public class A_QSort {
 
+public class C_QSortOptimized {
     private void quickSort(Segment[] segments, int left, int right) {
         if (left < right) {
-            int m = partition(segments, left, right);
-            quickSort(segments, left, m - 1);
-            quickSort(segments, m + 1, right);
+            int mid = (left + right) / 2;
+            int m1 = partition(segments, left, mid);
+            int m2 = partition(segments, m1 + 1, right);
+            quickSort(segments, left, m1 - 1);
+            quickSort(segments, m1 + 1,m2 );
+            quickSort(segments, m2 + 1, right);
         }
     }
     private int partition(Segment[] segments, int left, int right) {
@@ -60,7 +57,22 @@ public class A_QSort {
         segments[left] = tmp;
         return j;
     }
-
+    private int binarySearch(int pos, Segment[] segments){
+        // find in the arrays of segments
+        int left = 0;
+        int right = segments.length - 1;
+        while (left <= right){
+            int mid = left + (right - left) / 2;
+            if (pos < segments[mid].start){
+                right = mid - 1;
+            } else if (pos > segments[mid].stop){
+                left = mid + 1;
+            } else {
+                return mid;
+            }
+        }
+        return -1;
+    }
 
     //отрезок
     private class Segment  implements Comparable<Segment>{
@@ -75,11 +87,8 @@ public class A_QSort {
                 this.start = start;
                 this.stop = stop;
             }
-            //тут вообще-то лучше доделать конструктор на случай если
-            //концы отрезков придут в обратном порядке
         }
 
-        @Override
         public int compareTo(Segment o) {
             //подумайте, что должен возвращать компаратор отрезков
             if(this.start > o.start){
@@ -93,7 +102,7 @@ public class A_QSort {
     }
 
 
-    int[] getAccessory(InputStream stream) throws FileNotFoundException {
+    int[] getAccessory2(InputStream stream) throws FileNotFoundException {
         //подготовка к чтению данных
         Scanner scanner = new Scanner(stream);
         //!!!!!!!!!!!!!!!!!!!!!!!!!     НАЧАЛО ЗАДАЧИ     !!!!!!!!!!!!!!!!!!!!!!!!!
@@ -116,17 +125,26 @@ public class A_QSort {
         }
         //тут реализуйте логику задачи с применением быстрой сортировки
         //в классе отрезка Segment реализуйте нужный для этой задачи компаратор
-        quickSort(segments, 0, segments.length - 1);
-        for(int i = 0 ; i < m; i++){
-            int count = 0;
-            for(int j = 0; j < n; j++){
-                if(segments[j].start <= points[i] && segments[j].stop >= points[i]){
-                    count++;
-                }
-            }
-            result[i] = count;
-        }
 
+
+        quickSort(segments, 0, n-1);
+        for(int i = 0 ; i < m; i++){
+            int point1 = points[i];
+            int pos = binarySearch(point1, segments);
+            if(pos == -1){
+                result[i] = 0;
+            } else {
+                int count = 1;
+                for(int j = pos + 1; j < n; j++){
+                    if(segments[j].start <= point1 && segments[j].stop >= point1){
+                        count++;
+                    } else {
+                        break;
+                    }
+                }
+                result[i] = count;
+            }
+        }
 
         //!!!!!!!!!!!!!!!!!!!!!!!!!     КОНЕЦ ЗАДАЧИ     !!!!!!!!!!!!!!!!!!!!!!!!!
         return result;
@@ -135,9 +153,9 @@ public class A_QSort {
 
     public static void main(String[] args) throws FileNotFoundException {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/a_khmelev/lesson05/dataA.txt");
-        A_QSort instance = new A_QSort();
-        int[] result=instance.getAccessory(stream);
+        InputStream stream = new FileInputStream(root + "by/it/a_khmelev/lesson05/dataC.txt");
+        C_QSortOptimized instance = new C_QSortOptimized();
+        int[] result=instance.getAccessory2(stream);
         for (int index:result){
             System.out.print(index+" ");
         }
