+++ src/by/it/group310902/kasperets/lesson05/Lesson5Test.java
@@ -1,4 +1,4 @@
-package by.it.group310901.usov.lesson05;
+package by.it.group310902.kasperets.lesson05;
 
 import org.junit.Test;
