+++ src/by/it/group310902/kasperets/lesson05/dataA.txt
